feat(ee): move worker billing to ee/billing
